+++ winpr/libwinpr/utils/wlog/Message.c
@@ -54,9 +54,9 @@ char* WLog_Message_GetOutputFileName(int id, const char* ext)
 
 	ProcessId = GetCurrentProcessId();
 	if (id >= 0)
-		sprintf_s(FileName, 256, "%u-%d.%s", (unsigned int) ProcessId, id, ext);
+		sprintf_s(FileName, 256, "%"PRIu32"-%d.%s", ProcessId, id, ext);
 	else
-		sprintf_s(FileName, 256, "%u.%s", (unsigned int) ProcessId, ext);
+		sprintf_s(FileName, 256, "%"PRIu32".%s", ProcessId, ext);
 
 	FullFileName = GetCombinedPath(FilePath, FileName);
